# web/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MemEvolve API",
    "description": "Memory-enhanced proxy and evolution framework for OpenAI-compatible LLM APIs",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:11436",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "tags": ["Core API"],
        "summary": "Chat Completions",
        "description": "OpenAI-compatible chat completions with memory integration",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["messages"],
        "properties": {
          "model": { "type": "string" },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string",
                  "enum": ["system", "user", "assistant"]
                },
                "content": { "type": "string" }
              }
            }
          },
          "max_tokens": { "type": "integer" },
          "temperature": { "type": "number" },
          "stream": { "type": "boolean", "default": false }
        }
      },
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "object": { "type": "string" },
          "created": { "type": "integer" },
          "model": { "type": "string" },
          "choices": { "type": "array", "items": { "type": "object" } },
          "usage": { "type": "object" }
        }
      },
...
    }
  }
}